pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "researchmate"
version = "0.1.0"
description = "Multi-Step Research Agent with Tool Use for Academic Literature Analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["research", "agent", "langgraph", "llm", "academic", "literature-review"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langgraph>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "arxiv>=2.1.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "networkx>=3.0",
    "matplotlib>=3.9.0",
    "markdown>=3.7",
    "pandas>=2.2.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]
pdf = [
    "weasyprint>=62.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/researchmate"
Repository = "https://github.com/yourusername/researchmate"
Documentation = "https://github.com/yourusername/researchmate#readme"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
